backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-generativeai==0.8.3
pillow==11.0.0
chromadb==0.5.23
httpx==0.28.1
aiofiles==24.1.0
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
